Rename react-percy-storybook to percy-storybook
