+++ .gitignore
@@ -51,6 +51,3 @@ packages/*/lib
 
 # Packed npm
 *.tgz
-
-# Static storybook
-integration-tests/storybook-static
