+++ README.md
@@ -1,7 +1,7 @@
-# react-percy-storybook
-[![Package Status](https://img.shields.io/npm/v/@percy-io/react-percy-storybook.svg)](https://www.npmjs.com/package/@percy-io/react-percy-storybook)
-[![Build Status](https://travis-ci.org/percy/react-percy.svg?branch=master)](https://travis-ci.org/percy/react-percy)
+# percy-storybook
+[![Package Status](https://img.shields.io/npm/v/@percy-io/percy-storybook.svg)](https://www.npmjs.com/package/@percy-io/percy-storybook)
+[![Build Status](https://travis-ci.org/percy/percy-storybook.svg?branch=master)](https://travis-ci.org/percy/percy-storybook)
 
-React-percy-storybook adds [Percy](https://percy.io) visual testing and review to your [**React Storybook**](https://storybooks.js.org/).
+percy-storybook adds [Percy](https://percy.io) visual testing and review to your [**React Storybook**](https://storybooks.js.org/).
 
 #### Docs here: [https://percy.io/docs/clients/javascript/react-storybook](https://percy.io/docs/clients/javascript/react-storybook)
