+++ integration-tests/react-percy-storybook/package.json
@@ -21,7 +21,7 @@
   },
   "devDependencies": {
     "@percy-io/in-percy": "*",
-    "@percy-io/react-percy-storybook": "^1.1.5",
+    "@percy-io/percy-storybook": "^1.1.5",
     "@storybook/addon-actions": "^3.2.0",
     "@storybook/addon-info": "^3.2.4",
     "@storybook/addon-knobs": "^3.2.0",
