+++ packages/percy-storybook/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@percy-io/react-percy-storybook",
+  "name": "@percy-io/percy-storybook",
   "version": "1.1.5",
   "repository": {
     "type": "git",
@@ -41,5 +41,5 @@
       "window": true
     }
   },
-  "typings": "lib/react-percy-storybook.d.ts"
+  "typings": "lib/percy-storybook.d.ts"
 }
