+++ packages/percy-storybook/src/__tests__/getMinimumHeight-tests.js
