+++ packages/percy-storybook/src/cli.js
@@ -79,7 +79,7 @@ export async function run(argv) {
   const client = new ApiClient(
     process.env.PERCY_TOKEN,
     process.env.PERCY_API,
-    `react-percy-storybook/${VERSION}`,
+    `percy-storybook/${VERSION}`,
     `storybook/${storybookVersion()} react/${reactVersion()}`,
   );
