+++ packages/percy-storybook/src/getMinimumHeight.js
