+++ packages/percy-storybook/src/getStaticAssets.js
