+++ packages/percy-storybook/src/getWidths.js
